[feature] #41 Detail data tv / movie 나눔
