+++ domain/src/main/kotlin/com/meronacompany/domain/model/SpokenLanguage.kt
@@ -3,5 +3,5 @@ package com.meronacompany.domain.model
 data class SpokenLanguage(
     val english_name_: String,
     val iso_639_1_: String,
-    val name_: String
+    val name: String
 )
